messenger/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.0
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
decorator==5.1.1
Django==4.2.6
django-configurations==2.5
django-cors-headers==4.3.0
django-extensions==3.2.3
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.5
executing==2.0.0
frozenlist==1.4.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
ipython==8.16.1
jedi==0.19.1
matplotlib-inline==0.1.6
msgpack==1.0.7
multidict==6.0.4
mypy==1.6.0
mypy-extensions==1.0.0
onesignal-python-api==2.0.2
packaging==23.2
parso==0.8.3
pexpect==4.9.0
phonenumbers==8.13.22
pickleshare==0.7.5
Pillow==10.0.1
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyotp==2.9.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-django==4.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.11.2
twilio==8.9.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
yarl==1.9.2
zope.interface==6.1
